% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{doc_hot_keys}
\alias{doc_hot_keys}
\title{doc_hot_keys}
\usage{
doc_hot_keys(container, key, interval = NULL)
}
\arguments{
\item{container}{specify the container you wish to perform the action in}

\item{key}{press a key}

\item{interval}{#TODO: what is this?}
}
\description{
Press a hot key on your keyboard
}
