% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{doc_press}
\alias{doc_press}
\title{doc_press}
\usage{
doc_press(container, key)
}
\arguments{
\item{container}{specify the container you wish to perform the action in}

\item{key}{press a key}
}
\description{
Press a key on your keyboard
}
