% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{doc_press_tabs}
\alias{doc_press_tabs}
\title{doc_press_tabs}
\usage{
doc_press_tabs(container, n)
}
\arguments{
\item{container}{specify the container you wish to perform action in}

\item{n}{how often should the tab button be clicked}
}
\description{
Send several tabs (simplifies going through a website)
}
