% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{doc_type}
\alias{doc_type}
\title{doc_type}
\usage{
doc_type(container, message, interval = NULL)
}
\arguments{
\item{container}{specify the container you wish to perform the action in}

\item{message}{specify which text should be typed}

\item{interval}{#TODO: what is this?}
}
\description{
Type a (series of) text on your keyboard
}
